rag_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
chromadb==0.4.18
sentence-transformers==2.7.0
ollama==0.1.7
python-docx==1.1.0
PyPDF2==3.0.1
python-pptx==0.6.23
Pillow==10.1.0
numpy==1.24.3
aiofiles==24.1.0
